making some excellent progress
now ima make it so u can edit events. Ill make it so editting and deleting occurs from the day tab where u can change events etc and ill have it so untimed events are at the bottom and timed events are in the day and ill also make it os that there is an x button there too. ill also display the time on month events if there is one and possibly also display by time